+++ src/persisters/mod.rs
@@ -3,6 +3,3 @@
 pub mod db;
 pub mod fs;
 pub mod traits;
-
-pub use db::Orm;
-pub use fs::File;
